Update server functions
Add create listener for org update queue to push filtered changes to the root org model, and update listener on the root org model to push out filtered values to owner-readable/user-organizations.

Supports updating the name and photoURL for an org; though at the moment only name is modifiable in the interface.